chore: update hero image and add wedding details on the website
The hero image on the wedding website has been updated to a new image. Additionally, the wedding details section has been added to provide information about the event. The details include a welcome message, a note about the event being adults only, and the location of the wedding in Puerto Rico. The location section also includes a link to the address on Google Maps.